backend/requirements.txt
# FinOps Platform Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
email-validator==2.1.0
python-multipart==0.0.6
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
python-dateutil==2.8.2

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# Caching
redis==5.0.1
hiredis==2.2.3

# Background Tasks
celery==5.3.4
kombu==5.3.4
flower==2.0.1

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Cloud Provider SDKs
boto3==1.34.0
botocore==1.34.0
google-cloud-billing==1.12.1
azure-mgmt-consumption==10.0.0

# AI & Machine Learning
openai==1.6.1

# Utilities
pydantic-settings==2.1.0
python-dotenv==1.0.0
validators==0.22.0
psutil==5.9.6